requirements.txt
# Core
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0

# Web Search
tavily-python>=0.5.0

# Database
oracledb>=2.0.0

# OpenTelemetry
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp>=1.25.0
opentelemetry-instrumentation-requests>=0.46b0
opentelemetry-instrumentation-httpx>=0.46b0

# Frontend
streamlit>=1.32.0
streamlit-chat>=0.1.1
watchdog>=4.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
